search/mongodb/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
flask-pymongo==2.2.0
dnspython==1.16.0
bson==0.5.8
python-dotenv==0.10.1